chore: refactor for extensibility
